apiserver: replaced upgradingRoot using restrictRoot
